app/src/main/java/com/moviemagic/data/api/db/MovieDao.kt
package com.moviemagic.data.api.db

import androidx.room.Dao
import androidx.room.Insert
import androidx.room.OnConflictStrategy
import androidx.room.Query
import com.moviemagic.domain.entities.MovieEntity
import kotlinx.coroutines.flow.Flow

@Dao
interface MovieDao {

    @Query("SELECT * FROM movies LIMIT :limit OFFSET :page")
    fun getMovies(limit: Int, page: Int): Flow<List<MovieEntity>>

    @Insert(onConflict = OnConflictStrategy.REPLACE)
    suspend fun insertAll(movies: List<MovieEntity>)

    @Query("SELECT * FROM movies WHERE movies.id = :id")
    fun getMovieById(id: String): Flow<MovieEntity>

}